requirements.txt
python-dotenv==1.0.1
requests==2.32.3
python-telegram-bot==21.7
schedule==1.2.2
pytest==8.3.3
pytest-asyncio==0.24.0

